More Pcle cleanup
It turns out particles.lexc was not in use, I thus moved its two particle so particles_newwords.lexc and removed it. The 3 particles there no go to PCLE_.

All in all, the only-add-a-tag-lexca are not needed when we add tags in the stem file. Better would have been going to #.